@@ -18,7 +18,6 @@
 package license_test
 
 import (
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -83,7 +82,7 @@ func TestLicense(t *testing.T) {
 
 		switch filepath.Ext(path) {
 		case ".go":
-			src, err := ioutil.ReadFile(path)
+			src, err := os.ReadFile(path)
 			if err != nil {
 				return nil
 			}
@@ -96,7 +95,7 @@ func TestLicense(t *testing.T) {
 		case ".yaml":
 			fallthrough
 		case ".conf":
-			src, err := ioutil.ReadFile(path)
+			src, err := os.ReadFile(path)
 			if err != nil {
 				return nil
 			}
