@@ -19,6 +19,7 @@ package pulsar
 
 import (
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -27,8 +28,8 @@ func TestDefaultNackBackoffPolicy_Next(t *testing.T) {
 	defaultNackBackoff := new(defaultNackBackoffPolicy)
 
 	res0 := defaultNackBackoff.Next(0)
-	assert.Equal(t, int64(1000*30), res0)
+	assert.Equal(t, 1*time.Second, res0)
 
 	res5 := defaultNackBackoff.Next(5)
-	assert.Equal(t, int64(600000), res5)
+	assert.Equal(t, 32*time.Second, res5)
 }
